// frontend/src/app/core/services/api/model/scopeType.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Scope type for access control (internal, external, national)
 */
export const ScopeType = {
    Internal: 'INTERNAL',
    External: 'EXTERNAL',
    National: 'NATIONAL'
} as const;
export type ScopeType = typeof ScopeType[keyof typeof ScopeType];